+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Changed
 - Change method TSDataset.describe ([#409](https://github.com/tinkoff-ai/etna/pull/409))
 
+### Fixed
+- Add relevance_params in GaleShapleyFeatureSelectionTransform ([#410](https://github.com/tinkoff-ai/etna/pull/410))
+
 ## [1.5.0] - 2021-12-24
 ### Added
 - Holiday Transform ([#359](https://github.com/tinkoff-ai/etna/pull/359))